requirements.txt
Django==5.1
pillow==11.1.0
crispy-tailwind==1.0.3
djangorestframework==3.15.2
psycopg2==2.9.10